# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Accessible Games Platform API
  version: "1.0.0"
  description: >
    HTTP API for authentication, games catalog, user preferences, follows,
    reviews, and voice intent parsing used by the Team 13 project.
servers:
  - url: http://localhost:5000
    description: Local development
...
paths:
...
  /api/users/{id}/reviews:
    get:
      tags: [Users]
      summary: Get reviews created by a user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 10
      responses:
        "200":
          description: Reviews for user
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    rating:
                      type: integer
                    comment:
                      type: string
                    createdAt:
                      type: string
                      format: date-time
                    game:
                      type: object
                      nullable: true
                      properties:
                        id:
                          type: integer
                        title:
                          type: string
        "400":
          description: Invalid id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Fetch failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/users/{id}/accessibility-preferences:
    get:
      tags: [Users]
      summary: Get a user's accessibility preferences
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Preferences loaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessibilityPreferences"
        "400":
          description: Invalid id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Load failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      tags: [Users]
      summary: Update accessibility preferences
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccessibilityPreferences"
      responses:
        "200":
          description: Preferences updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessibilityPreferences"
        "400":
          description: Invalid id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Update failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/users/{id}/accessibility:
    put:
      tags: [Users]
      summary: Replace accessibility preferences (no auth enforced)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccessibilityPreferences"
      responses:
        "200":
          description: Preferences replaced
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  preferences:
                    $ref: "#/components/schemas/AccessibilityPreferences"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Update failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/users/{id}/follow/{gameId}:
    post:
      tags: [Users]
      summary: Follow a game
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Followed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  gameId:
                    type: integer
        "400":
          description: Invalid ids
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User or game not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Follow failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [Users]
      summary: Unfollow a game
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unfollowed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  gameId:
                    type: integer
        "400":
          description: Invalid ids
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User or game not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unfollow failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/users/{id}/followed-games:
    get:
      tags: [Users]
      summary: List games followed by a user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Followed games
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FollowedGame"
        "400":
          description: Invalid id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Load failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/users/{id}:
    patch:
      tags: [Users]
      summary: Update username or email
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                email:
                  type: string
                  format: email
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "400":
          description: Missing fields or invalid id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Update failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/users/{id}/password:
    patch:
      tags: [Users]
      summary: Change password
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword:
                  type: string
                  format: password
                newPassword:
                  type: string
                  format: password
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        "400":
          description: Missing fields or incorrect password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Change failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/users/{id}/recommended-games:
    get:
      tags: [Users]
      summary: Get recommended games based on accessibility preferences
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recommended games
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecommendedGame"
        "400":
          description: Invalid id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Fetch failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        error:
          type: string
        detail:
          type: string
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
      example:
        id: 12
        username: player1
        email: player1@example.com
        createdAt: "2024-05-01T12:34:56.000Z"
...
    AccessibilityPreferences:
      type: object
      properties:
        visual:
          type: boolean
        motor:
          type: boolean
        cognitive:
          type: boolean
        hearing:
          type: boolean
      example:
        visual: true
        motor: false
        cognitive: false
        hearing: true
    FollowedGame:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        images:
          type: array
          items:
            type: string
      example:
        id: 3
        title: Sprint Racer
        images: []
    RecommendedGame:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        platform:
          type: string
        rating:
          type: number
        images:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
      example:
        id: 5
        title: Calm Seas
        platform: Mobile
        rating: 4.2
        images:
          - /images/calm-seas-1.jpg
        tags: [Hearing, Captions]
...